Completed the implementation of 2D game manager with 3D game manager
Finished implementing the 2D combat in with your functions. Basically I did a coroutine for checking if the game is won or lost at the end. I made the player and enemy number static so that your gameManager can change those variables before startup. Once your system receives the int (1 for win, 0 for loss), then it will unload my scene and continue
so I am pretty sure my part is done
I put a slight delay so I have a IEnumerator method to yield the game for a slight bit before turning on the 3D camera